+++ tests/test_class_category.py
@@ -1,29 +1,30 @@
-from src.class_product import Product
+import pytest
 
 
-def test_product(first_product, second_product):
-    assert first_product.name == "Product"
-    assert first_product.description == "Description of the product"
-    assert first_product.price == 84.50
-    assert first_product.quantity == 10
+def test_category(first_category, second_category):
+    assert first_category.name == "Category"
+    assert first_category.description == "Description of the category"
+    assert (
+        first_category.products
+        == "Product, 84.5 руб. Остаток: 10 шт.\nProduct number two, 155.87 руб. Остаток: 34 шт.\n"
+    )
 
-    assert second_product.name == "Product number two"
-    assert second_product.description == "Description of the product number two"
-    assert second_product.price == 155.87
-    assert second_product.quantity == 34
+    assert first_category.category_count == 2
+    assert second_category.category_count == 2
 
-
-def test_new_product(product_dict):
-    product4 = Product.new_product(product_dict)
-    assert product4.name == "Product 4"
-    assert product4.description == "Description of the product 4"
-    assert product4.price == 145.75
-    assert product4.quantity == 23
+    assert first_category.product_count == 5
+    assert second_category.product_count == 5
 
 
-def test_prod_price_property(capsys, first_product):
-    first_product.price = -756.57
-    message = capsys.readouterr()
-    assert message.out.strip() == "Цена не должна быть нулевая или орицательная"
-    first_product.price = 756.57
-    assert first_product.price == 756.57
+def test_cat_products_property(first_category, second_category):
+    with pytest.raises(AttributeError):
+        print(first_category.__products)
+    assert (
+        first_category.products
+        == "Product, 84.5 руб. Остаток: 10 шт.\nProduct number two, 155.87 руб. Остаток: 34 шт.\n"
+    )
+    assert (
+        second_category.products
+        == "Product, 84.5 руб. Остаток: 10 шт.\nProduct number two, 155.87 руб. Остаток: 34 шт."
+        "\nProduct three, 8467.56 руб. Остаток: 32 шт.\n"
+    )
